pyproject.toml
[tool.poetry]
name = "preprocessing_sg"
version = "0.1.0"
description = ""
authors = ["EDF - F.Robin"]

[tool.poetry.dependencies]
python = "3.9.9"


[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
autopep8 = "1.6.0"
pyyaml = "6.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = [
    "tests",  # You should have a "tests" directory
]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"


[tool.autopep8]
max_line_length = 120
ignore = "E501,W6"  # or ["E501", "W6"]
in-place = true
recursive = true
aggressive = 3

